account-management-system/src/main/java/com/ams/accountmanagementsystem/model/User.java
package com.ams.accountmanagementsystem.model;

public class User {
    private String customerId;
    private String name;
    private String surname;

    public User(String customerId, String name, String surname) {
        this.customerId = customerId;
        this.name = name;
        this.surname = surname;
    }

    public String getcustomerId() {
        return customerId;
    }

    public void setcustomerId(String customerId) {
        this.customerId = customerId;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getSurname() {
        return surname;
    }

    public void setSurname(String surname) {
        this.surname = surname;
    }
}
